+++ app/src/main/java/com/anthonyla/paperize/feature/wallpaper/wallpaper_service/WallpaperService2.kt
@@ -17,6 +17,7 @@ import android.os.HandlerThread
 import android.os.IBinder
 import android.util.DisplayMetrics
 import android.util.Log
+import androidx.annotation.RequiresApi
 import androidx.core.app.NotificationCompat
 import androidx.core.net.toUri
 import com.anthonyla.paperize.R
@@ -37,6 +38,7 @@ import com.anthonyla.paperize.feature.wallpaper.presentation.MainActivity
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import java.io.IOException
@@ -119,6 +121,9 @@ class WallpaperService2: Service() {
         workerHandler.post {
             CoroutineScope(Dispatchers.IO).launch {
                 changeWallpaper(this@WallpaperService2, setHomeOrLock)
+                if (setHomeOrLock != null && !setHomeOrLock) {
+                    updateCurrentWallpaper(this@WallpaperService2, true)
+                }
             }
             stopSelf()
         }
@@ -138,10 +143,10 @@ class WallpaperService2: Service() {
         }
     }
 
-    private fun workerTaskUpdate() {
+    private fun workerTaskUpdate(setHomeOrLock: Boolean? = null) {
         workerHandler.post {
             CoroutineScope(Dispatchers.IO).launch {
-                updateCurrentWallpaper(this@WallpaperService2)
+                updateCurrentWallpaper(this@WallpaperService2, setHomeOrLock)
             }
             stopSelf()
         }
@@ -408,7 +413,7 @@ class WallpaperService2: Service() {
         }
     }
 
-    private suspend fun updateCurrentWallpaper(context: Context) {
+    private suspend fun updateCurrentWallpaper(context: Context, setHomeOrLock: Boolean? = null) {
         try {
             val selectedAlbum = selectedRepository.getSelectedAlbum().first().firstOrNull()
             if (selectedAlbum == null) {
@@ -422,10 +427,6 @@ class WallpaperService2: Service() {
                     onDestroy()
                     return
                 }
-                val formatter = DateTimeFormatter.ofLocalizedDateTime(FormatStyle.SHORT)
-                val time = LocalDateTime.now()
-                settingsDataStoreImpl.putString(SettingsConstants.LAST_SET_TIME, time.format(formatter))
-                settingsDataStoreImpl.putString(SettingsConstants.NEXT_SET_TIME, time.plusMinutes(timeInMinutes1.toLong()).format(formatter))
                 val scaling = settingsDataStoreImpl.getString(SettingsConstants.WALLPAPER_SCALING)?.let { ScalingConstants.valueOf(it) } ?: ScalingConstants.FILL
                 val darken = settingsDataStoreImpl.getBoolean(SettingsConstants.DARKEN) ?: false
                 val darkenPercentage = settingsDataStoreImpl.getInt(SettingsConstants.DARKEN_PERCENTAGE) ?: 100
@@ -435,36 +436,7 @@ class WallpaperService2: Service() {
                 selectedAlbum.let { it ->
                     val wallpaper1 = it.album.currentHomeWallpaper
                     val wallpaper2 = it.album.currentLockWallpaper
-                    if (wallpaper1 != null) {
-                        if (!setWallpaper(
-                                context = context,
-                                wallpaper = wallpaper1.toUri(),
-                                darken = darken,
-                                darkenPercent = darkenPercentage,
-                                scaling = scaling,
-                                setHome = setHome,
-                                setLock = setLock,
-                                setLockOrHome = true,
-                                blur = blur,
-                                blurPercent = blurPercentage
-                        )) {
-                            selectedAlbum.wallpapers.firstOrNull{ it.wallpaperUri == wallpaper1 }
-                                ?.let { it1 ->
-                                    selectedRepository.deleteWallpaper(it1)
-                                    selectedRepository.upsertSelectedAlbum(
-                                        it.copy(
-                                            album = it.album.copy(
-                                                homeWallpapersInQueue = it.album.homeWallpapersInQueue.filter { it != wallpaper1 },
-                                                lockWallpapersInQueue = it.album.lockWallpapersInQueue.filter { it != wallpaper1 }
-                                            ),
-                                            wallpapers = it.wallpapers.filter { it == it1 },
-                                        )
-                                    )
-                                    albumRepository.deleteWallpaper(it1)
-                                }
-                        }
-                    }
-                    if (wallpaper2 != null) {
+                    if (wallpaper2 != null && (setHomeOrLock == null || setHomeOrLock == false)) {
                         if (!setWallpaper(
                                 context = context,
                                 wallpaper = wallpaper2.toUri(),
@@ -477,20 +449,49 @@ class WallpaperService2: Service() {
                                 blur = blur,
                                 blurPercent = blurPercentage
                             )) {
-                            selectedAlbum.wallpapers.firstOrNull{ it.wallpaperUri == wallpaper2 }
-                                ?.let { it1 ->
-                                    selectedRepository.deleteWallpaper(it1)
-                                    selectedRepository.upsertSelectedAlbum(
-                                        it.copy(
-                                            album = it.album.copy(
-                                                homeWallpapersInQueue = it.album.homeWallpapersInQueue.filter { it != wallpaper1 },
-                                                lockWallpapersInQueue = it.album.lockWallpapersInQueue.filter { it != wallpaper1 }
-                                            ),
-                                            wallpapers = it.wallpapers.filter { it == it1 },
+                                selectedAlbum.wallpapers.firstOrNull{ it.wallpaperUri == wallpaper2 }
+                                    ?.let { it1 ->
+                                        selectedRepository.deleteWallpaper(it1)
+                                        selectedRepository.upsertSelectedAlbum(
+                                            it.copy(
+                                                album = it.album.copy(
+                                                    homeWallpapersInQueue = it.album.homeWallpapersInQueue.filter { it != wallpaper1 },
+                                                    lockWallpapersInQueue = it.album.lockWallpapersInQueue.filter { it != wallpaper1 }
+                                                ),
+                                                wallpapers = it.wallpapers.filter { it == it1 },
+                                            )
                                         )
-                                    )
-                                    albumRepository.deleteWallpaper(it1)
-                                }
+                                        albumRepository.deleteWallpaper(it1)
+                                    }
+                        }
+                    }
+                    if (wallpaper1 != null && (setHomeOrLock == null || setHomeOrLock == true)) {
+                        if (!setWallpaper(
+                                context = context,
+                                wallpaper = wallpaper1.toUri(),
+                                darken = darken,
+                                darkenPercent = darkenPercentage,
+                                scaling = scaling,
+                                setHome = setHome,
+                                setLock = setLock,
+                                setLockOrHome = true,
+                                blur = blur,
+                                blurPercent = blurPercentage
+                            )) {
+                                selectedAlbum.wallpapers.firstOrNull{ it.wallpaperUri == wallpaper1 }
+                                    ?.let { it1 ->
+                                        selectedRepository.deleteWallpaper(it1)
+                                        selectedRepository.upsertSelectedAlbum(
+                                            it.copy(
+                                                album = it.album.copy(
+                                                    homeWallpapersInQueue = it.album.homeWallpapersInQueue.filter { it != wallpaper1 },
+                                                    lockWallpapersInQueue = it.album.lockWallpapersInQueue.filter { it != wallpaper1 }
+                                                ),
+                                                wallpapers = it.wallpapers.filter { it == it1 },
+                                            )
+                                        )
+                                        albumRepository.deleteWallpaper(it1)
+                                    }
                         }
                     }
                 }
